+++ src/widgets/dock_area/mod.rs
@@ -1,33 +1,20 @@
+/// Due to there being a lot of code to show a dock in a ui every complementing
+/// method to ``show`` and ``show_inside`` is put in ``show_extra``.
+/// Otherwise ``mod.rs`` would be humongous.
+mod show;
+
+// Various components of the `DockArea` which is used when rendering
 mod allowed_splits;
 mod drag_and_drop;
 mod state;
-mod style_fade;
 mod tab_removal;
 
+use crate::{dock_state::DockState, NodeIndex, Style, SurfaceIndex, TabIndex};
 pub use allowed_splits::AllowedSplits;
-use std::{ops::RangeInclusive, time::Instant};
+use drag_and_drop::{DragData, HoverData};
 use tab_removal::TabRemoval;
 
-use crate::{
-    dock_state::DockState,
-    utils::{expand_to_pixel, map_to_pixel, rect_set_size_centered, rect_stroke_box},
-    widgets::popup::popup_under_widget,
-    Node, NodeIndex, OverlayType, Style, SurfaceIndex, TabAddAlign, TabDestination, TabIndex,
-    TabStyle, TabViewer,
-};
-
-use duplicate::duplicate;
-use egui::{
-    containers::*, emath::*, epaint::*, layers::*, Context, CursorIcon, Id, Layout, Response,
-    Sense, TextStyle, Ui, WidgetText,
-};
-use paste::paste;
-use state::State;
-
-use self::{
-    drag_and_drop::{DragData, HoverData, TreeComponent},
-    style_fade::{fade_dock_style, fade_visuals},
-};
+use egui::{emath::*, Id};
 
 /// Displays a [`DockState`] in `egui`.
 pub struct DockArea<'tree, Tab> {
@@ -41,6 +28,7 @@ pub struct DockArea<'tree, Tab> {
     draggable_tabs: bool,
     show_tab_name_on_hover: bool,
     scroll_area_in_tabs: bool,
+    show_window_heads: bool,
     allowed_splits: AllowedSplits,
     window_bounds: Option<Rect>,
 
@@ -76,6 +64,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
             new_focused: None,
             tab_hover_rect: None,
             window_bounds: None,
+            show_window_heads: true,
         }
     }
 
@@ -149,1332 +138,20 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         self
     }
 
-    /// Sets the bounds for any windows inside the [`DockArea`]. Defaults to the screen rect.
+    /// The bounds for any windows inside the [`DockArea`]. Defaults to the screen rect.
+    /// By default it's set to [`Context::screen_rect`].
     #[inline(always)]
     pub fn window_bounds(mut self, bounds: Rect) -> Self {
         self.window_bounds = Some(bounds);
         self
     }
-}
 
-// UI
-impl<'tree, Tab> DockArea<'tree, Tab> {
-    /// Show the `DockArea` at the top level.
-    ///
-    /// This is the same as doing:
-    ///
-    /// ```
-    /// # use egui_dock::{DockArea, DockState};
-    /// # use egui::{CentralPanel, Frame};
-    /// # struct TabViewer {}
-    /// # impl egui_dock::TabViewer for TabViewer {
-    /// #     type Tab = String;
-    /// #     fn ui(&mut self, ui: &mut egui::Ui, tab: &mut Self::Tab) {}
-    /// #     fn title(&mut self, tab: &mut Self::Tab) -> egui::WidgetText { (&*tab).into() }
-    /// # }
-    /// # let mut tree: DockState<String> = DockState::new(vec![]);
-    /// # let mut tab_viewer = TabViewer {};
-    /// # egui::__run_test_ctx(|ctx| {
-    /// CentralPanel::default()
-    ///     .frame(Frame::central_panel(&ctx.style()).inner_margin(0.))
-    ///     .show(ctx, |ui| {
-    ///         DockArea::new(&mut tree).show_inside(ui, &mut tab_viewer);
-    ///     });
-    /// # });
-    /// ```
-    ///
-    /// So you can't use the [`CentralPanel::show`] when using `DockArea`'s one.
-    ///
-    /// See also [`show_inside`](Self::show_inside).
-    #[inline]
-    pub fn show(self, ctx: &Context, tab_viewer: &mut impl TabViewer<Tab = Tab>) {
-        CentralPanel::default()
-            .frame(
-                Frame::central_panel(&ctx.style())
-                    .inner_margin(0.)
-                    .fill(Color32::TRANSPARENT),
-            )
-            .show(ctx, |ui| {
-                self.show_inside(ui, tab_viewer);
-            });
-    }
-
-    /// Shows the docking hierarchy inside a [`Ui`].
-    ///
-    /// See also [`show`](Self::show).
-    pub fn show_inside(mut self, ui: &mut Ui, tab_viewer: &mut impl TabViewer<Tab = Tab>) {
-        self.style
-            .get_or_insert(Style::from_egui(ui.style().as_ref()));
-        self.window_bounds.get_or_insert(ui.ctx().screen_rect());
-        let mut state = State::load(ui.ctx(), self.id);
-        let style = self.style.as_ref().unwrap();
-        let fade_surface =
-            self.hovered_window_surface(&mut state, style.overlay.feel.fade_hold_time, ui.ctx());
-        let fade_style = {
-            fade_surface.is_some().then(|| {
-                let mut fade_style = style.clone();
-                fade_dock_style(&mut fade_style, style.overlay.surface_fade_opacity);
-                (fade_style, style.overlay.surface_fade_opacity)
-            })
-        };
-
-        for &surface_index in self.dock_state.valid_surface_indices().iter() {
-            self.show_surface_inside(
-                surface_index,
-                ui,
-                tab_viewer,
-                &mut state,
-                fade_style.as_ref().map(|(style, factor)| {
-                    (style, *factor, fade_surface.unwrap_or(SurfaceIndex::main()))
-                }),
-            );
-        }
-
-        for index in self.to_remove.drain(..).rev() {
-            match index {
-                TabRemoval::Node(surface, node, tab) => {
-                    self.dock_state[surface].remove_tab((node, tab));
-                    if self.dock_state[surface].is_empty() && !surface.is_main() {
-                        self.dock_state.remove_surface(surface);
-                    }
-                }
-                TabRemoval::Window(index) => {
-                    self.dock_state.remove_surface(index);
-                }
-            }
-        }
-
-        for (surface_index, node_index, tab_index) in self.to_detach.drain(..).rev() {
-            let mouse_pos = ui.input(|input| input.pointer.hover_pos());
-            self.dock_state.detach_tab(
-                (surface_index, node_index, tab_index),
-                Rect::from_min_size(
-                    mouse_pos.unwrap_or(Pos2::ZERO),
-                    self.dock_state[surface_index][node_index]
-                        .rect()
-                        .map_or(Vec2::new(100., 150.), |rect| rect.size()),
-                ),
-            );
-        }
-
-        if let Some(focused) = self.new_focused {
-            self.dock_state.set_focused_node_and_surface(focused);
-        }
-
-        if let (Some(source), Some(hover)) = (self.drag_data.take(), self.hover_data.take()) {
-            let style = self.style.as_ref().unwrap();
-            state.set_drag_and_drop(source, hover, ui.ctx(), style);
-            let tab_dst = {
-                let layer_id = LayerId::new(Order::Foreground, "foreground".into());
-                ui.with_layer_id(layer_id, |ui| {
-                    self.show_drag_drop_overlay(ui, &mut state, tab_viewer)
-                })
-                .inner
-            };
-            if ui.input(|i| i.pointer.any_released()) {
-                let source = {
-                    match state.dnd.as_ref().unwrap().drag.src {
-                        TreeComponent::Tab(src_surf, src_node, src_tab) => {
-                            (src_surf, src_node, src_tab)
-                        }
-                        _ => todo!(
-                            "collections of tabs, like nodes and surfaces can't be docked (yet)"
-                        ),
-                    }
-                };
-                if let Some(destination) = tab_dst {
-                    self.dock_state.move_tab(source, destination);
-                    state.reset_drag();
-                }
-            }
-        }
-        state.store(ui.ctx(), self.id);
-    }
-
-    /// Show a single surface of a [`DockState`].
-    fn show_surface_inside(
-        &mut self,
-        surf_index: SurfaceIndex,
-        ui: &mut Ui,
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        state: &mut State,
-        fade_style: Option<(&Style, f32, SurfaceIndex)>,
-    ) {
-        if surf_index.is_main() {
-            self.show_root_surface_inside(ui, tab_viewer, state);
-        } else {
-            self.show_window_surface(ui, surf_index, tab_viewer, state, fade_style);
-        }
-    }
-
-    fn show_root_surface_inside(
-        &mut self,
-        ui: &mut Ui,
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        state: &mut State,
-    ) {
-        let surf_index = SurfaceIndex::main();
-
-        if self.dock_state.main_surface().is_empty() {
-            let rect = ui.available_rect_before_wrap();
-            let response = ui.allocate_rect(rect, Sense::hover());
-            if response.hovered() {
-                self.hover_data = Some(HoverData {
-                    rect,
-                    dst: TreeComponent::Surface(surf_index),
-                    tab: None,
-                })
-            }
-            return;
-        }
-
-        self.render_nodes(ui, tab_viewer, state, surf_index, None);
-    }
-
-    fn show_window_surface(
-        &mut self,
-        ui: &Ui,
-        surf_index: SurfaceIndex,
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        state: &mut State,
-        fade_style: Option<(&Style, f32, SurfaceIndex)>,
-    ) -> (Option<Pos2>, Option<DragData>) {
-        let title: WidgetText = format!("window {surf_index:?}").into();
-
-        // TODO: let the user dock entire windows,
-        // part 0 here goes into state.drag_start, 1 goes into self.drag_data
-        // from there implement move_window properly.
-        let window = {
-            let mut window_constructor = Window::new(title)
-                .title_bar(false)
-                .drag_bounds(self.window_bounds.unwrap());
-            let state = self.dock_state.get_window_state_mut(surf_index).unwrap();
-            if let Some(position) = state.next_position() {
-                window_constructor = window_constructor.current_pos(position);
-            }
-            if let Some(size) = state.next_size() {
-                window_constructor = window_constructor.fixed_size(size);
-            }
-            window_constructor
-        };
-
-        let (fade_factor, fade_style) = match fade_style {
-            Some((style, factor, surface_index)) => {
-                if surface_index == surf_index {
-                    (1.0, None)
-                } else {
-                    (factor, Some((style, factor)))
-                }
-            }
-            None => (1.0, None),
-        };
-        let mut frame = Frame::window(ui.style());
-        frame.fill = frame.fill.linear_multiply(fade_factor);
-        frame.stroke.color = frame.stroke.color.linear_multiply(fade_factor);
-        frame.shadow.color = frame.shadow.color.linear_multiply(fade_factor);
-        let response = window.frame(frame).show(ui.ctx(), |ui| {
-            ui.scope(|ui| {
-                if fade_factor != 1.0 {
-                    fade_visuals(ui.visuals_mut(), fade_factor);
-                }
-                self.render_nodes(ui, tab_viewer, state, surf_index, fade_style);
-            });
-        });
-
-        let screen_rect = match response {
-            Some(response) => response.response.rect,
-            None => Rect::NOTHING,
-        };
-
-        // This is our janky way of detecting drags on the window.
-        // `Some` indicates that a window was dragged, with just started specifying if this is the first frame of drag.
-        match self
-            .dock_state
-            .get_window_state_mut(surf_index)
-            .unwrap()
-            .was_dragged(ui.ctx(), screen_rect)
-        {
-            Some(just_started) => (
-                just_started.then_some(screen_rect.min),
-                Some(DragData {
-                    src: TreeComponent::Surface(surf_index),
-                    rect: self.dock_state[surf_index][NodeIndex::root()]
-                        .rect()
-                        .unwrap(),
-                }),
-            ),
-            None => (None, None),
-        }
-    }
-
-    fn render_nodes(
-        &mut self,
-        ui: &mut Ui,
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        state: &mut State,
-        surf_index: SurfaceIndex,
-        fade_style: Option<(&Style, f32)>,
-    ) {
-        // First compute all rect sizes in the node graph.
-        let max_rect = self.allocate_area_for_root_node(ui, surf_index);
-        for node_index in self.dock_state[surf_index].breadth_first_index_iter() {
-            if self.dock_state[surf_index][node_index].is_parent() {
-                self.compute_rect_sizes(ui, (surf_index, node_index), max_rect);
-            }
-        }
-
-        // Then, draw the bodies of each leaves.
-        for node_index in self.dock_state[surf_index].breadth_first_index_iter() {
-            if self.dock_state[surf_index][node_index].is_leaf() {
-                self.show_leaf(ui, state, (surf_index, node_index), tab_viewer, fade_style);
-            }
-        }
-
-        // Finally, draw separators so that their "interaction zone" is above
-        // bodies (see `SeparatorStyle::extra_interact_width`).
-        let fade_style = fade_style.map(|(style, _)| style);
-        for node_index in self.dock_state[surf_index].breadth_first_index_iter() {
-            if self.dock_state[surf_index][node_index].is_parent() {
-                self.show_separator(ui, (surf_index, node_index), fade_style);
-            }
-        }
-    }
-
-    fn allocate_area_for_root_node(&mut self, ui: &mut Ui, surface: SurfaceIndex) -> Rect {
-        let style = self.style.as_ref().unwrap();
-        let mut rect = ui.available_rect_before_wrap();
-
-        if let Some(margin) = style.dock_area_padding {
-            rect.min += margin.left_top();
-            rect.max -= margin.right_bottom();
-        }
-
-        ui.painter().rect_stroke(
-            rect,
-            style.main_surface_border_rounding,
-            style.main_surface_border_stroke,
-        );
-        if surface == SurfaceIndex::main() {
-            rect = rect.expand(-style.main_surface_border_stroke.width / 2.0);
-        }
-        ui.allocate_rect(rect, Sense::click());
-
-        if self.dock_state[surface].is_empty() {
-            return rect;
-        }
-        self.dock_state[surface][NodeIndex::root()].set_rect(rect);
-        rect
-    }
-
-    fn compute_rect_sizes(
-        &mut self,
-        ui: &Ui,
-        (surface_index, node_index): (SurfaceIndex, NodeIndex),
-        max_rect: Rect,
-    ) {
-        assert!(self.dock_state[surface_index][node_index].is_parent());
-
-        let style = self.style.as_ref().unwrap();
-        let pixels_per_point = ui.ctx().pixels_per_point();
-
-        duplicate! {
-            [
-                orientation   dim_point  dim_size  left_of    right_of;
-                [Horizontal]  [x]        [width]   [left_of]  [right_of];
-                [Vertical]    [y]        [height]  [above]    [below];
-            ]
-            if let Node::orientation { fraction, rect } = &mut self.dock_state[surface_index][node_index] {
-                debug_assert!(!rect.any_nan() && rect.is_finite());
-                let rect = expand_to_pixel(*rect, pixels_per_point);
-
-                let midpoint = rect.min.dim_point + rect.dim_size() * *fraction;
-                let left_separator_border = map_to_pixel(
-                    midpoint - style.separator.width * 0.5,
-                    pixels_per_point,
-                    f32::round
-                );
-                let right_separator_border = map_to_pixel(
-                    midpoint + style.separator.width * 0.5,
-                    pixels_per_point,
-                    f32::round
-                );
-
-                paste! {
-                    let left = rect.intersect(Rect::[<everything_ left_of>](left_separator_border)).intersect(max_rect);
-                    let right = rect.intersect(Rect::[<everything_ right_of>](right_separator_border)).intersect(max_rect);
-                }
-
-                self.dock_state[surface_index][node_index.left()].set_rect(left);
-                self.dock_state[surface_index][node_index.right()].set_rect(right);
-            }
-        }
-    }
-
-    fn show_separator(
-        &mut self,
-        ui: &mut Ui,
-        (surface_index, node_index): (SurfaceIndex, NodeIndex),
-        fade_style: Option<&Style>,
-    ) {
-        assert!(self.dock_state[surface_index][node_index].is_parent());
-
-        let style = fade_style.unwrap_or_else(|| self.style.as_ref().unwrap());
-        let pixels_per_point = ui.ctx().pixels_per_point();
-
-        duplicate! {
-            [
-                orientation   dim_point  dim_size;
-                [Horizontal]  [x]        [width];
-                [Vertical]    [y]        [height];
-            ]
-            if let Node::orientation { fraction, ref rect } = &mut self.dock_state[surface_index][node_index] {
-                let mut separator = *rect;
-
-                let midpoint = rect.min.dim_point + rect.dim_size() * *fraction;
-                separator.min.dim_point = midpoint - style.separator.width * 0.5;
-                separator.max.dim_point = midpoint + style.separator.width * 0.5;
-
-                let mut expand = Vec2::ZERO;
-                expand.dim_point += style.separator.extra_interact_width / 2.0;
-                let interact_rect = separator.expand2(expand);
-
-                let response = ui.allocate_rect(interact_rect, Sense::click_and_drag())
-                    .on_hover_and_drag_cursor(paste!{ CursorIcon::[<Resize orientation>]});
-
-                let midpoint = rect.min.dim_point + rect.dim_size() * *fraction;
-                separator.min.dim_point = map_to_pixel(
-                    midpoint - style.separator.width * 0.5,
-                    pixels_per_point,
-                    f32::round,
-                );
-                separator.max.dim_point = map_to_pixel(
-                    midpoint + style.separator.width * 0.5,
-                    pixels_per_point,
-                    f32::round,
-                );
-
-                let color = if response.dragged() {
-                    style.separator.color_dragged
-                } else if response.hovered() {
-                    style.separator.color_hovered
-                } else {
-                    style.separator.color_idle
-                };
-
-                ui.painter().rect_filled(separator, Rounding::none(), color);
-
-                // Update 'fraction' interaction after drawing separator,
-                // otherwise it may overlap on other separator / bodies when
-                // shrunk fast.
-                if let Some(pos) = response.interact_pointer_pos() {
-                    let dim_point = pos.dim_point;
-                    let delta = response.drag_delta().dim_point;
-
-                    if (delta > 0. && dim_point > midpoint && dim_point < rect.max.dim_point)
-                        || (delta < 0. && dim_point < midpoint && dim_point > rect.min.dim_point)
-                    {
-                        let range = rect.max.dim_point - rect.min.dim_point;
-                        let min = (style.separator.extra / range).min(1.0);
-                        let max = 1.0 - min;
-                        let (min, max) = (min.min(max), max.max(min));
-                        *fraction = (*fraction + delta / range).clamp(min, max);
-                    }
-                }
-
-                if response.double_clicked() {
-                    *fraction = 0.5;
-                }
-            }
-        }
-    }
-
-    fn show_leaf(
-        &mut self,
-        ui: &mut Ui,
-        state: &mut State,
-        (surface_index, node_index): (SurfaceIndex, NodeIndex),
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        fade_style: Option<(&Style, f32)>,
-    ) {
-        assert!(self.dock_state[surface_index][node_index].is_leaf());
-
-        let rect = self.dock_state[surface_index][node_index]
-            .rect()
-            .expect("This node must be a leaf");
-        let ui = &mut ui.child_ui_with_id_source(
-            rect,
-            Layout::top_down_justified(Align::Min),
-            (node_index, "node"),
-        );
-        let spacing = ui.spacing().item_spacing;
-        ui.spacing_mut().item_spacing = vec2(0.0, 0.0);
-        ui.set_clip_rect(rect);
-
-        let tabbar_response = self.tab_bar(
-            ui,
-            state,
-            (surface_index, node_index),
-            tab_viewer,
-            fade_style.map(|(style, _)| style),
-        );
-        self.tab_body(
-            ui,
-            state,
-            (surface_index, node_index),
-            tab_viewer,
-            spacing,
-            tabbar_response,
-            fade_style,
-        );
-
-        let tabs = self.dock_state[surface_index][node_index]
-            .tabs_mut()
-            .expect("This node must be a leaf here");
-        for (tab_index, tab) in tabs.iter_mut().enumerate() {
-            if tab_viewer.force_close(tab) {
-                self.to_remove
-                    .push((surface_index, node_index, TabIndex(tab_index)).into());
-            }
-        }
-    }
-
-    fn tab_bar(
-        &mut self,
-        ui: &mut Ui,
-        state: &mut State,
-        (surface_index, node_index): (SurfaceIndex, NodeIndex),
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        fade_style: Option<&Style>,
-    ) -> Response {
-        assert!(self.dock_state[surface_index][node_index].is_leaf());
-
-        let style = fade_style.unwrap_or_else(|| self.style.as_ref().unwrap());
-        let (tabbar_outer_rect, tabbar_response) = ui.allocate_exact_size(
-            vec2(ui.available_width(), style.tab_bar.height),
-            Sense::click(),
-        );
-        ui.painter().rect_filled(
-            tabbar_outer_rect,
-            style.tab_bar.rounding,
-            style.tab_bar.bg_fill,
-        );
-
-        let mut available_width = tabbar_outer_rect.width();
-
-        // Reserve space for the add button at the end of the tab bar.
-        if self.show_add_buttons {
-            available_width -= Style::TAB_ADD_BUTTON_SIZE;
-        }
-
-        let actual_width = {
-            let Node::Leaf { tabs, scroll, .. } = &mut self.dock_state[surface_index][node_index]
-            else {
-                unreachable!()
-            };
-
-            let tabbar_inner_rect = Rect::from_min_size(
-                (tabbar_outer_rect.min - pos2(-*scroll, 0.0)).to_pos2(),
-                vec2(tabbar_outer_rect.width(), tabbar_outer_rect.height()),
-            );
-
-            let tabs_ui = &mut ui.child_ui_with_id_source(
-                tabbar_inner_rect,
-                Layout::left_to_right(Align::Center),
-                "tabs",
-            );
-
-            let mut clip_rect = tabbar_outer_rect;
-            clip_rect.set_width(available_width);
-            tabs_ui.set_clip_rect(clip_rect);
-
-            // Desired size for tabs in "expanded" mode.
-            let prefered_width = style
-                .tab_bar
-                .fill_tab_bar
-                .then_some(available_width / (tabs.len() as f32));
-
-            self.tabs(
-                tabs_ui,
-                state,
-                (surface_index, node_index),
-                tab_viewer,
-                tabbar_outer_rect,
-                prefered_width,
-                fade_style,
-            );
-
-            // Draw hline from tab end to edge of tab bar.
-            let px = ui.ctx().pixels_per_point().recip();
-            let style = fade_style.unwrap_or_else(|| self.style.as_ref().unwrap());
-
-            ui.painter().hline(
-                tabs_ui.min_rect().right().min(clip_rect.right())..=tabbar_outer_rect.right(),
-                tabbar_outer_rect.bottom() - px,
-                (px, style.tab_bar.hline_color),
-            );
-
-            // Add button at the end of the tab bar.
-            if self.show_add_buttons {
-                let offset = match style.buttons.add_tab_align {
-                    TabAddAlign::Left => {
-                        (clip_rect.width() - tabs_ui.min_rect().width()).at_least(0.0)
-                    }
-                    TabAddAlign::Right => 0.0,
-                };
-                self.tab_plus(
-                    ui,
-                    node_index,
-                    tab_viewer,
-                    tabbar_outer_rect,
-                    offset,
-                    fade_style,
-                );
-            }
-
-            tabs_ui.min_rect().width()
-        };
-
-        self.tab_bar_scroll(
-            ui,
-            (surface_index, node_index),
-            actual_width,
-            available_width,
-            &tabbar_response,
-            fade_style,
-        );
-
-        tabbar_response
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn tabs(
-        &mut self,
-        tabs_ui: &mut Ui,
-        state: &mut State,
-        (surface_index, node_index): (SurfaceIndex, NodeIndex),
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        tabbar_outer_rect: Rect,
-        preferred_width: Option<f32>,
-        fade: Option<&Style>,
-    ) {
-        assert!(self.dock_state[surface_index][node_index].is_leaf());
-
-        let focused = self.dock_state.focused_leaf();
-        let tabs_len = {
-            let tabs = self.dock_state[surface_index][node_index]
-                .tabs()
-                .expect("This node must be a leaf here");
-            tabs.len()
-        };
-
-        for tab_index in 0..tabs_len {
-            let id = self
-                .id
-                .with((surface_index, "surface"))
-                .with((node_index, "node"))
-                .with((tab_index, "tab"));
-            let tab_index = TabIndex(tab_index);
-            let is_being_dragged =
-                tabs_ui.memory(|mem| mem.is_being_dragged(id)) && self.draggable_tabs;
-
-            if is_being_dragged {
-                tabs_ui.output_mut(|o| o.cursor_icon = CursorIcon::Grabbing);
-            }
-
-            let (is_active, label, tab_style, closeable) = {
-                let Node::Leaf { tabs, active, .. } =
-                    &mut self.dock_state[surface_index][node_index]
-                else {
-                    unreachable!()
-                };
-                let style = fade.unwrap_or_else(|| self.style.as_ref().unwrap());
-                let tab_style = tab_viewer.tab_style_override(&tabs[tab_index.0], &style.tab);
-                (
-                    *active == tab_index || is_being_dragged,
-                    tab_viewer.title(&mut tabs[tab_index.0]),
-                    tab_style.unwrap_or(style.tab.clone()),
-                    tab_viewer.closeable(&mut tabs[tab_index.0]),
-                )
-            };
-
-            let show_close_button = self.show_close_buttons && closeable;
-
-            let response = if is_being_dragged {
-                let layer_id = LayerId::new(Order::Tooltip, id);
-                let response = tabs_ui
-                    .with_layer_id(layer_id, |ui| {
-                        self.tab_title(
-                            ui,
-                            &tab_style,
-                            id,
-                            label,
-                            is_active && Some((surface_index, node_index)) == focused,
-                            is_active,
-                            is_being_dragged,
-                            preferred_width,
-                            show_close_button,
-                            fade,
-                        )
-                    })
-                    .response;
-
-                let sense = Sense::click_and_drag();
-                let response = tabs_ui.interact(response.rect, id, sense);
-
-                if let Some(pointer_pos) = tabs_ui.ctx().pointer_interact_pos() {
-                    let center = response.rect.center();
-                    let start = state.drag_start.unwrap_or(center);
-
-                    let delta = pointer_pos - start;
-                    if delta.x.abs() > 30.0 || delta.y.abs() > 6.0 {
-                        tabs_ui.ctx().translate_layer(layer_id, delta);
-
-                        self.drag_data = Some(DragData {
-                            src: TreeComponent::Tab(surface_index, node_index, tab_index),
-                            rect: self.dock_state[surface_index][node_index].rect().unwrap(),
-                        });
-                    }
-                }
-
-                response
-            } else {
-                let (mut response, close_response) = self.tab_title(
-                    tabs_ui,
-                    &tab_style,
-                    id,
-                    label,
-                    is_active && Some((surface_index, node_index)) == focused,
-                    is_active,
-                    is_being_dragged,
-                    preferred_width,
-                    show_close_button,
-                    fade,
-                );
-
-                let (close_hovered, close_clicked) = close_response
-                    .map(|res| (res.hovered(), res.clicked()))
-                    .unwrap_or_default();
-
-                let sense = if close_hovered {
-                    Sense::click()
-                } else {
-                    Sense::click_and_drag()
-                };
-
-                let is_lonely_tab = self.dock_state[surface_index].num_tabs() == 1;
-
-                let Node::Leaf { tabs, active, .. } =
-                    &mut self.dock_state[surface_index][node_index]
-                else {
-                    unreachable!()
-                };
-                let tab = &mut tabs[tab_index.0];
-                if self.show_tab_name_on_hover {
-                    response = response.on_hover_ui(|ui| {
-                        ui.label(tab_viewer.title(tab));
-                    });
-                }
-
-                if self.tab_context_menus {
-                    response = response.context_menu(|ui| {
-                        tab_viewer.context_menu(ui, tab, node_index);
-                        if (surface_index.is_main() || !is_lonely_tab)
-                            && ui.button("Eject").clicked()
-                        {
-                            self.to_detach.push((surface_index, node_index, tab_index));
-                            ui.close_menu();
-                        }
-                        if show_close_button && ui.button("Close").clicked() {
-                            if tab_viewer.on_close(tab) {
-                                self.to_remove
-                                    .push((surface_index, node_index, tab_index).into());
-                            } else {
-                                *active = tab_index;
-                                self.new_focused = Some((surface_index, node_index));
-                            }
-                            ui.close_menu();
-                        }
-                    });
-                }
-
-                if close_clicked {
-                    if tab_viewer.on_close(tab) {
-                        self.to_remove
-                            .push((surface_index, node_index, tab_index).into());
-                    } else {
-                        *active = tab_index;
-                        self.new_focused = Some((surface_index, node_index));
-                    }
-                }
-                let response = tabs_ui.interact(response.rect, id, sense);
-                if response.drag_started() {
-                    state.drag_start = response.hover_pos();
-                }
-
-                if let Some(pos) = tabs_ui.input(|i| i.pointer.hover_pos()) {
-                    // Use response.rect.contains instead of
-                    // response.hovered as the dragged tab covers
-                    // the underlying tab
-                    if state.drag_start.is_some() && response.rect.contains(pos) {
-                        self.tab_hover_rect = Some((response.rect, tab_index));
-                    }
-                }
-
-                response
-            };
-
-            // Paint hline below each tab unless its active (or option says otherwise).
-            let Node::Leaf { tabs, active, .. } = &mut self.dock_state[surface_index][node_index]
-            else {
-                unreachable!()
-            };
-            let tab = &mut tabs[tab_index.0];
-            let style = match fade {
-                Some(fade) => fade,
-                None => self.style.as_ref().unwrap(),
-            };
-            let tab_style = tab_viewer.tab_style_override(tab, &style.tab);
-            let tab_style = tab_style.as_ref().unwrap_or(&style.tab);
-
-            if !is_active || tab_style.hline_below_active_tab_name {
-                let px = tabs_ui.ctx().pixels_per_point().recip();
-                tabs_ui.painter().hline(
-                    response.rect.x_range(),
-                    tabbar_outer_rect.bottom() - px,
-                    (px, style.tab_bar.hline_color),
-                );
-            }
-
-            if response.clicked() {
-                *active = tab_index;
-                self.new_focused = Some((surface_index, node_index));
-            }
-
-            if self.show_close_buttons && response.middle_clicked() {
-                if tab_viewer.on_close(tab) {
-                    self.to_remove
-                        .push((surface_index, node_index, tab_index).into());
-                } else {
-                    *active = tab_index;
-                    self.new_focused = Some((surface_index, node_index));
-                }
-            }
-
-            tab_viewer.on_tab_button(tab, &response);
-        }
-    }
-
-    fn tab_plus(
-        &mut self,
-        ui: &mut Ui,
-        node_index: NodeIndex,
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        tabbar_outer_rect: Rect,
-        offset: f32,
-        fade_style: Option<&Style>,
-    ) {
-        let rect = Rect::from_min_max(
-            tabbar_outer_rect.right_top() - vec2(Style::TAB_ADD_BUTTON_SIZE + offset, 0.0),
-            tabbar_outer_rect.right_bottom() - vec2(offset, 2.0),
-        );
-
-        let ui = &mut ui.child_ui_with_id_source(
-            rect,
-            Layout::left_to_right(Align::Center),
-            (node_index, "tab_add"),
-        );
-
-        let (rect, mut response) = ui.allocate_exact_size(ui.available_size(), Sense::click());
-
-        response = response.on_hover_cursor(CursorIcon::PointingHand);
-
-        let style = fade_style.unwrap_or_else(|| self.style.as_ref().unwrap());
-        let color = if response.hovered() {
-            ui.painter()
-                .rect_filled(rect, Rounding::none(), style.buttons.add_tab_bg_fill);
-            style.buttons.add_tab_active_color
-        } else {
-            style.buttons.add_tab_color
-        };
-
-        let mut plus_rect = rect;
-
-        rect_set_size_centered(&mut plus_rect, Vec2::splat(Style::TAB_ADD_PLUS_SIZE));
-
-        ui.painter().line_segment(
-            [plus_rect.center_top(), plus_rect.center_bottom()],
-            Stroke::new(1.0, color),
-        );
-        ui.painter().line_segment(
-            [plus_rect.right_center(), plus_rect.left_center()],
-            Stroke::new(1.0, color),
-        );
-
-        // Draw button left border.
-        ui.painter().vline(
-            rect.left(),
-            rect.y_range(),
-            Stroke::new(
-                ui.ctx().pixels_per_point().recip(),
-                style.buttons.add_tab_border_color,
-            ),
-        );
-
-        let popup_id = ui.id().with("tab_add_popup");
-        popup_under_widget(ui, popup_id, &response, |ui| {
-            tab_viewer.add_popup(ui, node_index);
-        });
-
-        if response.clicked() {
-            if self.show_add_popup {
-                ui.memory_mut(|mem| mem.toggle_popup(popup_id));
-            }
-            tab_viewer.on_add(node_index);
-        }
-    }
-
-    /// * `active` means "the tab that is opened in the parent panel".
-    /// * `focused` means "the tab that was last interacted with".
-    ///
-    /// Returns the main button response plus the response of the close button, if any.
-    #[allow(clippy::too_many_arguments)]
-    fn tab_title(
-        &mut self,
-        ui: &mut Ui,
-        tab_style: &TabStyle,
-        id: Id,
-        label: WidgetText,
-        focused: bool,
-        active: bool,
-        is_being_dragged: bool,
-        prefered_width: Option<f32>,
-        show_close_button: bool,
-        fade: Option<&Style>,
-    ) -> (Response, Option<Response>) {
-        let style = fade.unwrap_or_else(|| self.style.as_ref().unwrap());
-        let galley = label.into_galley(ui, None, f32::INFINITY, TextStyle::Button);
-        let x_spacing = 8.0;
-        let text_width = galley.size().x + 2.0 * x_spacing;
-        let close_button_size = if show_close_button {
-            Style::TAB_CLOSE_BUTTON_SIZE.min(style.tab_bar.height)
-        } else {
-            0.0
-        };
-
-        // Compute total width of the tab bar.
-        let minimum_width = tab_style
-            .minimum_width
-            .unwrap_or(0.0)
-            .at_least(text_width + close_button_size);
-        let tab_width = prefered_width.unwrap_or(0.0).at_least(minimum_width);
-
-        let (rect, mut response) =
-            ui.allocate_exact_size(vec2(tab_width, ui.available_height()), Sense::hover());
-        if !ui.memory(|mem| mem.is_anything_being_dragged()) && self.draggable_tabs {
-            response = response.on_hover_cursor(CursorIcon::PointingHand);
-        }
-
-        let tab_style = if focused || is_being_dragged {
-            &tab_style.focused
-        } else if active {
-            &tab_style.active
-        } else if response.hovered() {
-            &tab_style.hovered
-        } else {
-            &tab_style.inactive
-        };
-
-        // Draw the full tab first and then the stroke on top to avoid the stroke
-        // mixing with the background color.
-        ui.painter()
-            .rect_filled(rect, tab_style.rounding, tab_style.bg_fill);
-        let stroke_rect = rect_stroke_box(rect, 1.0);
-        ui.painter().rect_stroke(
-            stroke_rect,
-            tab_style.rounding,
-            Stroke::new(1.0, tab_style.outline_color),
-        );
-        if !is_being_dragged {
-            // Make the tab name area connect with the tab ui area.
-            ui.painter().hline(
-                RangeInclusive::new(
-                    stroke_rect.min.x + f32::max(tab_style.rounding.sw, 1.5),
-                    stroke_rect.max.x - f32::max(tab_style.rounding.se, 1.5),
-                ),
-                stroke_rect.bottom(),
-                Stroke::new(2.0, tab_style.bg_fill),
-            );
-        }
-
-        let mut text_rect = rect;
-        text_rect.set_width(tab_width - close_button_size);
-
-        let text_pos = {
-            let pos = Align2::CENTER_CENTER.pos_in_rect(&text_rect.shrink2(vec2(x_spacing, 0.0)));
-            pos - galley.size() / 2.0
-        };
-
-        let override_text_color = (!galley.galley_has_color).then_some(tab_style.text_color);
-
-        ui.painter().add(TextShape {
-            pos: text_pos,
-            galley: galley.galley,
-            underline: Stroke::NONE,
-            override_text_color,
-            angle: 0.0,
-        });
-
-        let close_response = show_close_button.then(|| {
-            let mut close_button_rect = rect;
-            close_button_rect.set_left(text_rect.right());
-            close_button_rect =
-                Rect::from_center_size(close_button_rect.center(), Vec2::splat(close_button_size));
-
-            let response = ui
-                .interact(close_button_rect, id, Sense::click())
-                .on_hover_cursor(CursorIcon::PointingHand);
-
-            let color = if response.hovered() {
-                style.buttons.close_tab_active_color
-            } else {
-                style.buttons.close_tab_color
-            };
-
-            if response.hovered() {
-                let mut rounding = tab_style.rounding;
-                rounding.nw = 0.0;
-                rounding.sw = 0.0;
-
-                ui.painter().rect_filled(
-                    close_button_rect,
-                    rounding,
-                    style.buttons.add_tab_bg_fill,
-                );
-            }
-
-            let mut x_rect = close_button_rect;
-            rect_set_size_centered(&mut x_rect, Vec2::splat(Style::TAB_CLOSE_X_SIZE));
-            ui.painter().line_segment(
-                [x_rect.left_top(), x_rect.right_bottom()],
-                Stroke::new(1.0, color),
-            );
-            ui.painter().line_segment(
-                [x_rect.right_top(), x_rect.left_bottom()],
-                Stroke::new(1.0, color),
-            );
-
-            response
-        });
-
-        (response, close_response)
-    }
-
-    fn tab_bar_scroll(
-        &mut self,
-        ui: &mut Ui,
-        (surface_index, node_index): (SurfaceIndex, NodeIndex),
-        actual_width: f32,
-        available_width: f32,
-        tabbar_response: &Response,
-        fade_style: Option<&Style>,
-    ) {
-        let Node::Leaf { scroll, .. } = &mut self.dock_state[surface_index][node_index] else {
-            unreachable!()
-        };
-        let overflow = (actual_width - available_width).at_least(0.0);
-        let style = fade_style.unwrap_or_else(|| self.style.as_ref().unwrap());
-
-        // Compare to 1.0 and not 0.0 to avoid drawing a scroll bar due
-        // to floating point precision issue during tab drawing.
-        if overflow > 1.0 {
-            if style.tab_bar.show_scroll_bar_on_overflow {
-                // Draw scroll bar
-                let bar_height = 7.5;
-                let (scroll_bar_rect, _scroll_bar_response) = ui.allocate_exact_size(
-                    vec2(ui.available_width(), bar_height),
-                    Sense::click_and_drag(),
-                );
-
-                // Compute scroll bar handle position and size.
-                let overflow_ratio = actual_width / available_width;
-                let scroll_ratio = -*scroll / overflow;
-
-                let scroll_bar_handle_size = overflow_ratio.recip() * scroll_bar_rect.width();
-                let scroll_bar_handle_start = lerp(
-                    scroll_bar_rect.left()..=scroll_bar_rect.right() - scroll_bar_handle_size,
-                    scroll_ratio,
-                );
-                let scroll_bar_handle_rect = Rect::from_min_size(
-                    pos2(scroll_bar_handle_start, scroll_bar_rect.min.y),
-                    vec2(scroll_bar_handle_size, bar_height),
-                );
-
-                let scroll_bar_handle_response = ui.interact(
-                    scroll_bar_handle_rect,
-                    self.id.with((node_index, "node")),
-                    Sense::drag(),
-                );
-
-                // Coefficient to apply to input displacements so that we move the scroll by the correct amount.
-                let points_to_scroll_coefficient =
-                    overflow / (scroll_bar_rect.width() - scroll_bar_handle_size);
-
-                *scroll -= scroll_bar_handle_response.drag_delta().x * points_to_scroll_coefficient;
-
-                if let Some(pos) = ui.input(|i| i.pointer.hover_pos()) {
-                    if scroll_bar_rect.contains(pos) {
-                        *scroll += ui.input(|i| i.scroll_delta.y + i.scroll_delta.x)
-                            * points_to_scroll_coefficient;
-                    }
-                }
-
-                // Draw the bar.
-                ui.painter()
-                    .rect_filled(scroll_bar_rect, 0.0, ui.visuals().extreme_bg_color);
-
-                ui.painter().rect_filled(
-                    scroll_bar_handle_rect,
-                    bar_height / 2.0,
-                    ui.visuals()
-                        .widgets
-                        .style(&scroll_bar_handle_response)
-                        .bg_fill,
-                );
-            }
-
-            // Handle user input.
-            if tabbar_response.hovered() {
-                *scroll += ui.input(|i| i.scroll_delta.y + i.scroll_delta.x);
-            }
-        }
-
-        *scroll = scroll.clamp(-overflow, 0.0);
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn tab_body(
-        &mut self,
-        ui: &mut Ui,
-        state: &State,
-        (surface_index, node_index): (SurfaceIndex, NodeIndex),
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        spacing: Vec2,
-        tabbar_response: Response,
-        fade: Option<(&Style, f32)>,
-    ) {
-        let (body_rect, _body_response) =
-            ui.allocate_exact_size(ui.available_size_before_wrap(), Sense::hover());
-
-        let Node::Leaf {
-            rect,
-            tabs,
-            active,
-            viewport,
-            ..
-        } = &mut self.dock_state[surface_index][node_index]
-        else {
-            unreachable!();
-        };
-
-        if let Some(tab) = tabs.get_mut(active.0) {
-            *viewport = body_rect;
-
-            if ui.input(|i| i.pointer.any_click()) {
-                if let Some(pos) = ui.input(|i| i.pointer.hover_pos()) {
-                    if body_rect.contains(pos) && Some(ui.layer_id()) == ui.ctx().layer_id_at(pos) {
-                        self.new_focused = Some((surface_index, node_index));
-                    }
-                }
-            }
-
-            let (style, fade_factor) = fade.unwrap_or_else(|| (self.style.as_ref().unwrap(), 1.0));
-            let tabs_styles = tab_viewer.tab_style_override(tab, &style.tab);
-
-            let tabs_style = tabs_styles.as_ref().unwrap_or(&style.tab);
-
-            if tab_viewer.clear_background(tab) {
-                ui.painter()
-                    .rect_filled(body_rect, 0.0, tabs_style.tab_body.bg_fill);
-            }
-
-            // Construct a new ui with the correct tab id.
-            //
-            // We are forced to use `Ui::new` because other methods (eg: push_id) always mix
-            // the provided id with their own which would cause tabs to change id when moved
-            // from node to node.
-            let id = self.id.with(tab_viewer.id(tab));
-            ui.ctx().check_for_id_clash(id, body_rect, "a tab with id");
-            let ui = &mut Ui::new(
-                ui.ctx().clone(),
-                ui.layer_id(),
-                id,
-                body_rect,
-                ui.clip_rect(),
-            );
-            ui.set_clip_rect(Rect::from_min_max(ui.cursor().min, ui.clip_rect().max));
-
-            // Use initial spacing for ui.
-            ui.spacing_mut().item_spacing = spacing;
-
-            // Offset the background rectangle up to hide the top border behind the clip rect.
-            // To avoid anti-aliasing lines when the stroke width is not divisible by two, we
-            // need to calculate the effective anti-aliased stroke width.
-            let effective_stroke_width = (tabs_style.tab_body.stroke.width / 2.0).ceil() * 2.0;
-            let tab_body_rect = Rect::from_min_max(
-                ui.clip_rect().min - vec2(0.0, effective_stroke_width),
-                ui.clip_rect().max,
-            );
-            ui.painter().rect(
-                rect_stroke_box(tab_body_rect, tabs_style.tab_body.stroke.width),
-                tabs_style.tab_body.rounding,
-                tabs_style.tab_body.bg_fill,
-                tabs_style.tab_body.stroke,
-            );
-
-            if self.scroll_area_in_tabs {
-                ScrollArea::both().show(ui, |ui| {
-                    Frame::none()
-                        .inner_margin(tabs_style.tab_body.inner_margin)
-                        .show(ui, |ui| {
-                            if fade_factor != 1.0 {
-                                fade_visuals(ui.visuals_mut(), fade_factor);
-                            }
-                            let available_rect = ui.available_rect_before_wrap();
-                            ui.expand_to_include_rect(available_rect);
-                            tab_viewer.ui(ui, tab);
-                        });
-                });
-            } else {
-                Frame::none()
-                    .inner_margin(tabs_style.tab_body.inner_margin)
-                    .show(ui, |ui| {
-                        if fade_factor != 1.0 {
-                            fade_visuals(ui.visuals_mut(), fade_factor);
-                        }
-                        tab_viewer.ui(ui, tab);
-                    });
-            }
-        }
-
-        if let Some(pointer) = ui.input(|i| i.pointer.hover_pos()) {
-            // Prevent borrow checker issues.
-            let rect = rect.to_owned();
-
-            // Use rect.contains instead of response.hovered as the dragged tab covers
-            // the underlying responses.
-            if state.drag_start.is_some() && rect.contains(pointer) {
-                let on_title_bar = tabbar_response.rect.contains(pointer);
-                let (dst, tab) = {
-                    match self.tab_hover_rect {
-                        Some((rect, tab_index)) => (
-                            TreeComponent::Tab(surface_index, node_index, tab_index),
-                            Some(rect),
-                        ),
-                        None => (
-                            TreeComponent::Node(surface_index, node_index),
-                            on_title_bar.then_some(tabbar_response.rect),
-                        ),
-                    }
-                };
-
-                self.hover_data = Some(HoverData { rect, dst, tab });
-            }
-        }
-    }
-
-    /// Returns some when windows are fading, and what surface index is being hovered over
+    /// Enables or disables the collapsing header and close button on windows.
+    /// By default it's `true`.
     #[inline(always)]
-    fn hovered_window_surface(
-        &self,
-        state: &mut State,
-        hold_time: f32,
-        ctx: &Context,
-    ) -> Option<SurfaceIndex> {
-        if let Some(dnd_state) = &state.dnd {
-            if dnd_state.is_locked(self.style.as_ref().unwrap(), ctx) {
-                state.window_fade = Some((Instant::now(), dnd_state.hover.dst.surface_address()));
-            }
-        }
-
-        state.window_fade.and_then(|(time, surface)| {
-            ctx.request_repaint();
-            (time.elapsed().as_secs_f32() < hold_time).then_some(surface)
-        })
-    }
-
-    /// Resolve where a dragged tab would land given it's dropped this frame, returns `None` when the resulting drop is an invalid move.
-    fn show_drag_drop_overlay(
-        &mut self,
-        ui: &Ui,
-        state: &mut State,
-        tab_viewer: &impl TabViewer<Tab = Tab>,
-    ) -> Option<TabDestination> {
-        let drag_state = state.dnd.as_mut().unwrap();
-        let style = self.style.as_ref().unwrap();
-
-        // If were hovering over ourselves, we're not moving anywhere.
-        if drag_state.hover.dst.node_address() == drag_state.drag.src.node_address()
-            && drag_state.is_on_title_bar()
-        {
-            return None;
-        }
-
-        let deserted_node = {
-            match (
-                drag_state.drag.src.node_address(),
-                drag_state.hover.dst.node_address(),
-            ) {
-                ((src_surf, Some(src_node)), (dst_surf, Some(dst_node))) => {
-                    src_surf == dst_surf
-                        && src_node == dst_node
-                        && self.dock_state[src_surf][src_node].tabs_count() == 1
-                }
-                _ => false,
-            }
-        };
-
-        // Not all scenarios can house all splits.
-        let restricted_splits = if drag_state.hover.dst.is_surface() || deserted_node {
-            AllowedSplits::None
-        } else {
-            AllowedSplits::All
-        };
-        let allowed_splits = self.allowed_splits & restricted_splits;
-
-        let allowed_in_window = match drag_state.drag.src {
-            TreeComponent::Tab(surface, node, tab) => {
-                let Node::Leaf { tabs, .. } = &mut self.dock_state[surface][node] else {
-                    unreachable!("tab drags can only come from leaf nodes")
-                };
-                tab_viewer.allowed_in_windows(&mut tabs[tab.0])
-            }
-            _ => todo!("collections of tabs, like nodes or surfaces, can't be dragged! (yet)"),
-        };
-
-        if let Some(pointer) = ui.input(|i| i.pointer.hover_pos()) {
-            drag_state.pointer = pointer;
-        }
-
-        let window_bounds = self.window_bounds.unwrap();
-        if drag_state.is_on_title_bar()
-            || style.overlay.overlay_type == OverlayType::HighlightedAreas
-        {
-            drag_state.resolve_traditional(
-                ui,
-                style,
-                allowed_splits,
-                allowed_in_window,
-                window_bounds,
-            )
-        } else {
-            drag_state.resolve_icon_based(
-                ui,
-                style,
-                allowed_splits,
-                allowed_in_window,
-                window_bounds,
-            )
-        }
+    pub fn show_window_heads(mut self, show_window_heads: bool) -> Self {
+        self.show_window_heads = show_window_heads;
+        self
     }
 }
